code refactored

+++ src/main/java/com/bank/usermanagement/dto/TransactionDto.java
@@ -2,7 +2,7 @@ package com.bank.usermanagement.dto;
 
 import lombok.*;
 
-import java.math.BigDecimal;
+
 
 @Getter
 @Setter
@@ -12,6 +12,6 @@ import java.math.BigDecimal;
 public class TransactionDto {
     private Long fromAccountId;
     private Long toAccountId;
-    private BigDecimal amount;
+    private double amount;
     private String Status;
 }
