+++ src/main/java/com/bank/usermanagement/dto/TransactionRequest.java
@@ -2,11 +2,11 @@ package com.bank.usermanagement.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
+
 
 @Data
 public class TransactionRequest {
     private Long accountId;
-    private BigDecimal amount;
+    private double amount;
     private String transactionPassword;
 }
